Make species details appear in the same fragment when selected from the search results.
Switch the search fragment to use the same adapter as the group
fragment. The old adapter can be deleted now (although its name is
probably better, so do it the other way around instead.. also we should
use the indexer stuff from the old one too).

Add functionality to style the title and subtitle associated with each
fragment. Don't actually use this yet since the italicisation looks a
bit silly, and I'm not sure there's value putting the scientific name
there (it's shown a little lower down anyway). In fact, what I mean is
that not only do we not use the new styling functionality, we no longer
show the scientific name.