Splash screen.
Also make onCreate of the DB run on the calling thread -- we were
calling it from a non-main thread anyway, and this way we know when it's
done.

Hide the progress spinner and text. See what Rachel thinks, but I don't
think they add a great deal.

Remove the artifical 1s splash screen before showing the main activity.

Remove the DB progress methods.