Fix splash screen for M devices.
The mountains were hiding behind the bottom navigation bar. Change the
theme to set the background rather than window background to fix this.

There was also a bit of a jank when we changed from the fullscreen
splash screen to the non-fullscreen main screen, so make the splash
activity non-fullscreen.

Remove the activity_splash layout, and just leave the theme set instead.