Add support for up + back instead of just back=up.
Code needs some tidying, but the interactions actually seem pretty nice!
Idea is that each fragment has a "level" and a "slaveness" (at the
moment this is represented by odd vs even levels). When you push a
fragment, if its level is the same as the level of the current leaf then
it stays at the same hierarchy level, and either replaces the whole
screen (if it's a master) or just the right (if it's a slave). Back
goes back in time. Up goes back until the level changes. Note that the
actual numerical ordering of the levels is irrelevant.