Updates to FragmentController.
Abstract away the title/subtitle stuff into a new class FragmentInfo.
FragmentController now isn't supposed to care about the title/subtitle
of the fragment, it just gets a FragmentInfo blob to use. It's now up to
MainActivity to set the title and up button and so on appropriately.
FragmentController is just meant to organise which fragments get shown
when.

FragmentInfo also includes an ID which identifies a fragment to the
level of "looks the same". This allows us to make sure the same fragment
is never pushed twice in a row. This lives in FragmentController now for
convenience, even though it really belongs in MainActivity.